% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iDESC_function.R
\name{iDESC}
\alias{iDESC}
\title{iDESC: main function for DE analysis using zero-inflated negative binomial mixed model}
\usage{
iDESC(
  mat,
  meta,
  subject_var,
  group_var,
  zp_group_var,
  norm_opt = c("SeqDepth", "SizeFactor", "User", "None"),
  user_sf = NULL,
  sub_cell_filtering = 5,
  gene_sub_filtering = 1,
  gene_cell_filtering = 0.05,
  zp.thresh = 0.05,
  diagnosis = F
)
}
\arguments{
\item{mat}{Count matrix}

\item{meta}{Data frame including information for cells}

\item{subject_var}{The name of subject information in meta}

\item{group_var}{The name of group/disease information for DE analysis in meta}

\item{zp_group_var}{The name of group/disease information for differences of inflated zeros in meta}

\item{norm_opt}{Option for normalizing factors}

\item{user_sf}{Option for user-specific normalizing factors}

\item{sub_cell_filtering}{Filtering on cells within each subject}

\item{gene_sub_filtering}{Filtering on genes based on expression in each subject}

\item{gene_cell_filtering}{Filtering on genes based on expression across all cells}

\item{zp.thresh}{Threshold of zero proportions}

\item{diagnosis}{Option for doing diagnosis on zero proportions}
}
\value{
A result table contains parameter estimation
}
\description{
iDESC: main function for DE analysis using zero-inflated negative binomial mixed model
}
