% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IPF_example}
\alias{IPF_example}
\title{Single cell RNA-seq data for IPF patients and healthy controls.}
\format{
A list with a count matrix and data frame:
\describe{
\item{mat}{Count matrix for 20 genes and 1489 cells}
\item{meta}{Disease, subject, sequencing depth information for 1489 cells}
...
}
}
\source{
\url{http://www.ipfcellatlas.com/}
}
\usage{
IPF_example
}
\description{
A data set containing 20 genes and 1489 cells from IPF patients and healthy controls
}
\keyword{datasets}
