% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization_factors.R
\name{normalization_factors}
\alias{normalization_factors}
\title{normalization_factors: Calculating different types of normalization factors}
\usage{
normalization_factors(mat, norm_opt, user_sf)
}
\arguments{
\item{mat}{Count matrix}

\item{norm_opt}{Option for normalizing factors}

\item{user_sf}{Option for user-specific normalizing factors}
}
\value{
A vector of normalizing factors
}
\description{
normalization_factors: Calculating different types of normalization factors
}
