% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocessing}
\alias{preprocessing}
\title{preprocessing: filtering step on genes/cells}
\usage{
preprocessing(
  mat,
  meta,
  subject_var,
  group_var,
  sub_cell_filtering,
  gene_sub_filtering,
  gene_cell_filtering
)
}
\arguments{
\item{mat}{Count matrix}

\item{meta}{Data frame including information for cells}

\item{subject_var}{The name of subject information in meta}

\item{group_var}{The name of group/disease information for DE analysis in meta}

\item{sub_cell_filtering}{Filtering on cells within each subject}

\item{gene_sub_filtering}{Filtering on genes based on expression in each subject}

\item{gene_cell_filtering}{Filtering on genes based on expression across all cells}
}
\value{
A list of processed count matrix and data frame
}
\description{
preprocessing: filtering step on genes/cells
}
