% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zp_diagnosis.R
\name{zp_diagnosis}
\alias{zp_diagnosis}
\title{zp_diagnosis: Diagnosis about group differences on inflated zeros}
\usage{
zp_diagnosis(vec, meta, zp_group_var, zp.thresh = zp.thresh)
}
\arguments{
\item{vec}{A vector of data (one gene) that need diagnosis}

\item{meta}{Data frame including information for cells}

\item{zp_group_var}{The name of group/disease information for differences of inflated zeros in meta}

\item{zp.thresh}{Threshold of zero proportions}
}
\value{
A string with the format "Two Pi Model" or "One Pi Model"
}
\description{
zp_diagnosis: Diagnosis about group differences on inflated zeros
}
