% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zp_prediction.R
\name{zp_prediction}
\alias{zp_prediction}
\title{zp_prediction: Predict dropout rate from LOESS curve}
\usage{
zp_prediction(mat, norm_factor, span)
}
\arguments{
\item{mat}{Count matrix}

\item{norm_factor}{normalization factors from function: "normalization_factors"}

\item{span}{smoothing parameter}
}
\value{
A vector of LOESS predicted dropout rate
}
\description{
zp_prediction: Predict dropout rate from LOESS curve
}
