% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iDESC_function.R
\name{iDESC}
\alias{iDESC}
\title{iDESC: main function for DE analysis using zero-inflated negative binomial mixed model}
\usage{
iDESC(
  mat,
  meta,
  subject_var,
  group_var,
  norm_opt = c("SeqDepth", "SizeFactor", "User", "None"),
  user_sf = NULL,
  sub_cell_filtering = 5,
  gene_sub_filtering = 0,
  gene_cell_filtering = 0.05,
  ncell_filtering = 1,
  span = 0.05,
  loess_control = "",
  covariates = NULL,
  cores = 1
)
}
\arguments{
\item{mat}{Count matrix}

\item{meta}{Data frame including information for cells}

\item{subject_var}{The name of subject information in meta}

\item{group_var}{The name of group/disease information for DE analysis in meta}

\item{norm_opt}{Option for normalizing factors}

\item{user_sf}{Option for user-specific normalizing factors}

\item{sub_cell_filtering}{Filtering on cells within each subject}

\item{gene_sub_filtering}{Filtering on genes based on expression in each subject}

\item{gene_cell_filtering}{Filtering on genes based on expression across all cells}

\item{ncell_filtering}{Filtering on cells based on the number of genes expressed}

\item{span}{smoothing parameter for LOESS curve}

\item{loess_control}{Optional. If set to "control", the LOESS smoothing function will use
\code{loess.control(surface = "direct")}. This setting is recommended when working with sparse or
zero-heavy datasets, where the default LOESS method may struggle to define neighborhoods for
smoothing. Using "direct" computation ensures more stable and accurate smoothing but may be
slower. Activate this option if you encounter warnings like "k-d tree limited by memory" or
if the LOESS fit appears unstable.}

\item{covariates}{Optional. A vector of covariates to include in the model.}

\item{cores}{Number of cores for parallel processing (default is 1)}
}
\value{
A list containing:
\itemize{
\item{\strong{model_results}: A data frame containing the results of the differential expression analysis. For each gene, this includes estimated effect sizes (coefficients), dispersion parameters, zero-inflation estimates, statistical significance (raw p-values and FDR-corrected p-values), and model deviances.}
\item{\strong{problematic_genes}: A data frame listing genes that were excluded from the main analysis because they were expressed in cells from only one group. For each gene, the table reports the number of cells with non-zero expression in each group. This can be used for further diagnostic or exploratory analyses, such as assessing group-specific expression patterns.}
}
}
\description{
iDESC: main function for DE analysis using zero-inflated negative binomial mixed model
}
\examples{

library(iDESC)
data(IPF_example)
mat=IPF_example$mat
meta=IPF_example$meta
sequencing_depth=IPF_example$meta$sequencing_depth
result=iDESC(mat,meta,subject_var="subject",group_var="disease",norm_opt="User",user_sf = sequencing_depth,span = 0.7,cores = 4)
}
