% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iDESC_function.R
\name{iDESC}
\alias{iDESC}
\title{iDESC: main function for DE analysis using zero-inflated negative binomial mixed model}
\usage{
iDESC(
  mat,
  meta,
  subject_var,
  group_var,
  norm_opt = c("SeqDepth", "SizeFactor", "User", "None"),
  user_sf = NULL,
  sub_cell_filtering = 5,
  gene_sub_filtering = 0,
  gene_cell_filtering = 0.05,
  ncell_filtering = 1,
  span = 0.05
)
}
\arguments{
\item{mat}{Count matrix}

\item{meta}{Data frame including information for cells}

\item{subject_var}{The name of subject information in meta}

\item{group_var}{The name of group/disease information for DE analysis in meta}

\item{norm_opt}{Option for normalizing factors}

\item{user_sf}{Option for user-specific normalizing factors}

\item{sub_cell_filtering}{Filtering on cells within each subject}

\item{gene_sub_filtering}{Filtering on genes based on expression in each subject}

\item{gene_cell_filtering}{Filtering on genes based on expression across all cells}

\item{ncell_filtering}{Filtering on cells based on the number of genes expressed}

\item{span}{smoothing parameter for LOESS curve}
}
\value{
A result table contains parameter estimation
}
\description{
iDESC: main function for DE analysis using zero-inflated negative binomial mixed model
}
\examples{

library(iDESC)
data(IPF_example)
mat=IPF_example$mat
meta=IPF_example$meta
sequencing_depth=IPF_example$meta$sequencing_depth
result=iDESC(mat,meta,subject_var="subject",group_var="disease",norm_opt="User",user_sf = sequencing_depth,span = 0.7)
}
